django-serverless-aws/adocato/requirements.txt
argcomplete==3.1.6
asgiref==3.7.2
boto3==1.29.6
botocore==1.32.6
certifi==2023.11.17
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
configparser==6.0.0
Django==4.2.7
django-s3-storage==0.15.0
djangorestframework==3.14.0
durationpy==0.5
hjson==3.1.0
idna==3.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
placebo==0.9.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
troposphere==4.5.2
typing_extensions==4.8.0
urllib3==1.26.18
Werkzeug==3.0.1
zappa==0.58.0
